sla/requirements.txt
Django==1.5.2
django-extensions
south
django-debug-toolbar
requests
python-dateutil<2.0
